add password to and change url of bookings page
